:bug: fix: 테스트 오류 해결
