+++ src/test/java/com/mallang/mallang_backend/domain/payment/service/request/PaymentRedisServiceTest.java
@@ -1,5 +1,6 @@
 package com.mallang.mallang_backend.domain.payment.service.request;
 
+import com.mallang.mallang_backend.domain.payment.repository.PaymentRepository;
 import com.mallang.mallang_backend.global.exception.ServiceException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,6 +20,7 @@ import static com.mallang.mallang_backend.global.constants.AppConstants.IDEM_KEY
 import static com.mallang.mallang_backend.global.constants.AppConstants.ORDER_ID_PREFIX;
 import static com.mallang.mallang_backend.global.exception.ErrorCode.ORDER_ID_CONFLICT;
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
@@ -30,6 +32,9 @@ class PaymentRedisServiceTest {
 
     @MockitoBean
     private RedisTemplate<String, Object> redisTemplate;
+
+    @MockitoBean
+    private PaymentRepository paymentRepository;
 
     @MockitoBean
     private ValueOperations<String, Object> valueOperations;
@@ -80,6 +85,25 @@ class PaymentRedisServiceTest {
         // Verify: 1초기 시도 + 4회 재시도 = 총 5회 호출
         verify(valueOperations, times(5)).set(
                 anyString(), anyString(), any(Duration.class));
+    }
+
+    @Test
+    void Redis_저장_실패시_DB_미반영_검증() {
+        // given: Redis 모킹 설정
+        ValueOperations<String, Object> valueOps = mock(ValueOperations.class);
+
+        when(redisTemplate.opsForValue()).thenReturn(valueOps);
+        when(redisTemplate.hasKey(anyString())).thenReturn(false);
+
+        doThrow(new QueryTimeoutException("Connection failed"))
+                .when(valueOps).set(anyString(), anyString(), any(Duration.class));
+
+        // when
+        assertThrows(ServiceException.class, () ->
+                paymentRedisService.saveDataToRedis("token123", "order456", 10000));
+
+        // then: DB에 주문 정보 미존재 검증
+        assertFalse(paymentRepository.existsByOrderId("order456"));
     }
 
     @Test
